Created .sql file
Realized that the hand in must be in a .sql file so I redid the lab
making sure to keep SELECT, FROM, and WHERE on their own lines for easy
grading. I saved the file directly from the query editor and named the
file Lab3.sql and placed it in the Lab3 folder. I also deleted all the
screenshots from the previous commit.