Lab5 Upload
Refreshed the cap2 database for this assignment just in case. Question 2
was the toughest. I could not get the duplicate pids of the 3 agents,
only the pid of the 3 respective agents would return. #7 was not bad,
but for some reason it took me the longest. It was one of those times
when you just really over think something and make it way more
complicated than it had to be. Came back after a break and solved it in
5 minutes...